requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-decouple==3.8
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
